added option to remove TR fraction
Substituted --removeTrackRefs option by --keepTrackRefsFraction <percentage>
which enforces removal of TrackRefs.root in all but <percentage> fraction of jobs
Decision is taken randomly according to ALIEN_PROC_IS assigned in the sub-job JDL.
In absence of externally provided ALIEN_PROC_ID a warning will be printed
and locally generated random number will be assigned.

Default value of TrackRefs to keep is set in the declaration
CONFIG_KEEPTRACKREFSFRACTION=100, can be changed to any value between 0 and 100.
This patch should be accompanied by wildcarding e.g. as TrackRefs*.root saving request
in the JDL