Changes for ROOT6
The macros have now been split in 2. Suppose that the original macro
was called "macro.C", and here you had loading of external macros
and execution of functions defined there. Now you have:

- macro.C --> same name as before; here you load external macros
and call the main macro that does the actual stuff (name is unchanged
since the scripts will call this macro - so we don't have to
change the scripts)

- main_macro.C --> "main_" added in front; here we do the actual stuff;
this macro is called by the one above

May god be with us... :)