Add generator for Herwig7
Events are generated with the Herwig7
executable in an external process and
read in AlRoot with AliGenReaderHepMC.
Config files for Herwig are handled
as template files replacing instances
of CONFIG* with the corresponding
environment variables set by dpgsim.sh.
Translation of the templates into
Matchbox files is done using JINJA.
Initially a template is provided for
Min. Bias pp, more templates eventually
to be added in the future.