Fix NULL vs. "" handling in PHOS CaloCells macros (#161)
Task `AliAnalysisTaskCaloCellsQA` exists with two different signatures. We need
to know which one we are using. Note that checking the AliRoot version is weak
(since we may not have a variable holding the version, for instance during
development). We now:

* check the macro file for the presence of the old signature,
* define a preprocessor directive accordingly (via `gInterpreter`).

The directive is processed before we invoke the corresponding
`main_QAtrain_duo.C` macro, and expanded properly there, in both ROOT 5 and 6.

Note that the old strategies are wrong/faulty for different reasons:

* conditionally checking the value of a boolean _at runtime_ does not prevent
  ROOT 6 from still processing the lines it won't actually execute, leading to
  errors for the incorrect signatures,
* loading the correct signature based on the ROOT version is wrong, since the
  change has been made in AliPhysics and it was already in use with the new
  signature for AliRoot ROOT 5-based versions