[ALIROOT-8678] Set weight power to 4
Test productions for pp 13 TeV show that the weight power
6 is too high and biases the spectrum too much towards high
pt, reducing the statistics at low pt. As the goal of a weighted
production is to have the spectrum as flat in pt-hard as possible
the weight needs to be reduced. Setting weight power to 4
based on PYTHIA8 tutorial main08.cc.