Config for Xic0 production in Pb-Pb (AliGenParams)
Config file for simulation of Pb-Pb collisions with Xic0. Xic0 are injected directly with AliGenParams and decayed with the Pythia6 decayer. Note that Pythia6 has a wrong lifetime for Xic0, an external decay table is needed to fix it. 
If MC-to-MC embedding, 50 Xic0 + 50 Xic0bar are injected in central collisions up to b=5 fm, then 15+15. 
If there is not embedding, it is assumed that the production is for central collisions and 50+50 Xic0 are injected